Add `ApiConfig` and `BaseWorker` for enhanced v2 API and worker abstraction
- Introduced `ApiConfig` for parsing, validation, and translation of v2 API configs.
- Added `BaseWorker` abstract class supporting dual-write latency tracking, shutdown operations, and messaging pattern handling.
- Improved structure for modularity, shared state, and concurrency.

+++ kubemq-java/src/main/java/io/kubemq/sdk/pubsub/PubSubStats.java
@@ -21,10 +21,10 @@ import lombok.Setter;
 public class PubSubStats {
 
   /** The number of messages for the Pub/Sub channel. */
-  private int messages;
+  private long messages;
 
   /** The volume of data (in bytes) for the Pub/Sub channel. */
-  private int volume;
+  private long volume;
 
   /** The number of messages currently waiting to be processed by the Pub/Sub channel. */
   private int waiting;
